+++ Assets/Scripts/Yang/PlayerMovement.cs.meta
@@ -1,6 +1,6 @@
 fileFormatVersion: 2
-guid: 0e7f86d1b1f7a57428033dd679b77114
-timeCreated: 1470201542
+guid: aac49fcdc549a224899de4f64750abee
+timeCreated: 1470244385
 licenseType: Free
 MonoImporter:
   serializedVersion: 2
