+++ Assets/Scripts/Yang/backups/ArmRotation.cs.meta
@@ -1,6 +1,6 @@
 fileFormatVersion: 2
-guid: b33bcdbea535adf42be4aa75ace6e3f5
-timeCreated: 1470023640
+guid: 7651f4ed91bc1d4459386bc1d5252fed
+timeCreated: 1470244385
 licenseType: Free
 MonoImporter:
   serializedVersion: 2
