+++ Assets/Scripts/Yang/backups/CameraFollow.cs.meta
@@ -1,6 +1,6 @@
 fileFormatVersion: 2
-guid: 3e699d0e3881daa4b969cdcb95db8b82
-timeCreated: 1470065537
+guid: 6da107dddfc4e1b4f9836c120497259e
+timeCreated: 1470244385
 licenseType: Free
 MonoImporter:
   serializedVersion: 2
