+++ Assets/Scripts/Yang/backups/DeathTrigger.cs
