+++ Assets/Scripts/Yang/backups/DeathTrigger.cs.meta
@@ -1,6 +1,6 @@
 fileFormatVersion: 2
-guid: ab29359c61dc8b74392f9d14de886a4b
-timeCreated: 1470011469
+guid: b95d9e696b7a6644ab17bdd13d1a8778
+timeCreated: 1470244385
 licenseType: Free
 MonoImporter:
   serializedVersion: 2
