+++ Assets/Scripts/Yang/backups/Killable.cs
