+++ Assets/Scripts/Yang/backups/Killable.cs.meta
@@ -1,6 +1,6 @@
 fileFormatVersion: 2
-guid: 6229d3798376255429d93e122d5df255
-timeCreated: 1470016143
+guid: 330886d79f2ac474e98e9d362aef1148
+timeCreated: 1470244385
 licenseType: Free
 MonoImporter:
   serializedVersion: 2
