+++ Assets/Scripts/Yang/backups/LaserController.cs
