+++ Assets/Scripts/Yang/backups/LaserController.cs.meta
@@ -1,6 +1,6 @@
 fileFormatVersion: 2
-guid: 2ec8efab170fadf48bbd12a2c8554814
-timeCreated: 1470011865
+guid: e9392d18ed460b34ca9d521be049cc00
+timeCreated: 1470244385
 licenseType: Free
 MonoImporter:
   serializedVersion: 2
