+++ Assets/Scripts/Yang/backups/SeerShoot.cs.meta
@@ -1,6 +1,6 @@
 fileFormatVersion: 2
-guid: 052de4158076b1b4fb1c5f385326e02d
-timeCreated: 1470011468
+guid: c87ddc3d473a1394591f522e52a0ef1d
+timeCreated: 1470244385
 licenseType: Free
 MonoImporter:
   serializedVersion: 2
